\name{ProteinLength.Gene}
\alias{ProteinLength.Gene}
\alias{ProteinLength}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Determine Protein Length from CDS
}
\description{
This function will return the length of a protein from a given CDS of a gene.
}
\usage{
\method{ProteinLength}{Gene}(object, ...)

ProteinLength(object)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{
A list that contains a character string called CDS.
}
\item{...}{
Additional arguments
}
}

