\name{proteinLength}
\alias{proteinLength}
\alias{proteinLength.Gene}

\title{
Protein Length
}
\description{
A function that Calculates the length in Amino acids of the CDS sequece
}
\usage{
proteinLength(object, ...)
\method{proteinLength}{Gene}(object, ...)
}

\arguments{
  \item{object}{
A list generated by the gene function
}
  \item{\dots}{
other arguments 
}
}
\value{
A number indiccating amino acid length
}

\examples{
myobject <- Gene("Abc000011", "XYZ", c(8020, 112, 900),
               "ATGAAATAG")
proteinLength(myobject) 

}
