\name{print.Gene}
\alias{print.Gene}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Print Gene ID, Symbol, and Protein Length
}
\description{
This function will print out the gene ID, gene symbol, and length of a protein encoded by the gene.
}
\usage{
\method{print}{Gene}(object, ...)

print.Gene(object)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{
A list that contains character strings called ID, symbol, and CDS.
}
}

